kiss

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: jetownle <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/30 17:40:59 by jetownle          #+#    #+#             */
-/*   Updated: 2019/05/15 16:52:25 by jetownle         ###   ########.fr       */
+/*   Updated: 2019/05/16 21:33:32 by jetownle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
